# src/main/resources/openapi3_1.yaml
openapi: 3.1.0
info:
    title: "Clinica"
    summary: "Applicacion para una clinica generica"
    description: "En esta aplicacion se desarrollara compoennetes y modulos de software que se crean necesarios dentro del contexto de un dominio 'Clinica' generico"
    contact:
        name: Bautista
        email: Batubasilio@hotmail.com
    version: 1.0.0
    license: 
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
    - url: "http://localhost:{port}/api/{basePath}"
      variables:
        port:
            enum:
                - '8080'
                - '8081'
            default: '8080'
        basePath:
            default: v1
...
components:
    schemas:
        UserDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                username:
                    type: string
                    descrtiption: Nombre de usuario
                password:
                    type: string
                    maxLength: 10
                    minLength: 6
                    examples: [mpkJordan]
                role:
                    type: string
                    examples: [ADMIN,USER]
                    default: USER
                    enum:
                        - ADMIN
                        - USER
                status:
                    type: string
                    examples: [ENABLE, DISABLE, SUSPENDED]
                    default: ENABLE
                    enum:
                        - ENABLE
                        - DISABLE
                        - SUSPENDED
        
                dateCreated:
                    type: string
                    format: date
                dateDeleted:
                    type: string
                    format: date
            required:
                - id
                - username
                - password
                - role
                - status
                - dateCreated
        UserRequestDTO:
            type: object
            properties:
                username:
                    type: string
                    description: Nombre de usuario
                password:
                    type: string
                    maxLength: 10
                    minLength: 6
                    examples: [mpkJORNDAN]
                role:
                    type: string
                    examples: [ADMIN,USER]
                    default: USER
                    enum:
                        - ADMIN
                        - USER
                status:
                    type: string
                    examples: [ ENABLE, DISABLE, SUSPENDED ]
                    default: ENABLE
                    enum:
                        - ENABLE
                        - DISABLE
                        - SUSPENDED
                dateCreated:
                    type: string
                    format: date
                dateDeleted:
                    type: string
                    format: date
            required: 
              - username
              - password
              - role
              - dateCreated
        UserListDTO:
            type: object
            properties:
                items:
                    type: array
                    items:
                       $ref: '#/components/schemas/UserDTO'
...
        BadRequestErrorDTO:
            type: object
            properties:
                error:
                    type: string
                    examples:   [Bad request] 
                message:
                    type: string
                    examples: [La peticion no es valida, verificar los parametros de peticiones] 
        NotFoundMessageDTO:
            type: object
            properties:
                error:
                    type: string
                    examples:
                        - NotFound
                message:
                    type: string
                    examples:
                        - The resources is not avalible
        InternalErrorDTO:
            type: object
            properties:
                error:
                    type: string
                    examples:
                        - InternalServerError
                message:
                    type: string
                    examples:
                        - An error occurred on the server
        UnAuthorizeRequestDTO:
            type: object
            properties:
                error:
                    type: string
                    examples:
                        - UnAuthorizeRequest
                message:
                    type: string
                    examples:
                        - UnAuthorized User to make a request
paths:
    /users/userCreate:
        post:
            tags:
                - User
            security:
                - bearerAuth: []
            summary: Crear usuario
            description: Este end point ejecuta las rutinas de creacion de un usuario, con ncesidad de que el usuario se encuentre en el rango de administrador
            operationId: createUser
            requestBody:
                description: Created user object
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserRequestDTO'
            responses:
                '200':
                    description: Peticion coreccta
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDTO'
                404:
                    description: "Si la peticon falla"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                401:
                    description: "Si el usuario no tien autorizacion"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/UnAuthorizeRequestDTO'
                400:
                    description: 'Si la peticion no es correcta'
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/BadRequestErrorDTO'
                500:
                    description: "Error intreno del servidor"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/InternalErrorDTO'      
    /users/findAll:
        get:
            tags:
                - User
            security:
                - bearerAuth: []
            summary: Retorna el listado de todos los usuarios
            description: "Retorna el listado de todos los usuarios, Autenticacion requerida"  
            operationId: retriveAllUser
            responses:
                200:
                    description: Si la peticion es exitosa
                    content: 
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserListDTO'
                                    
                400:
                    description: Si el endpoint no existe
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                404:
                    description: "Si la peticon falla"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                401:
                    description: "Si el usuario no tien autorizacion"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/UnAuthorizeRequestDTO'
                500:
                    description: "Error intreno del servidor"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/InternalErrorDTO'     
    /users/findUser/{userId}:
        get:
            tags:
                - User
            security:
                - bearerAuth: []
            summary: Retorna un usuario
            description: Retorna un usuario el cual su id corresponda con el ingresado por URL
            operationId: retriveUser
            parameters:
                - name: userId 
                  in: path
                  description: el ID del usuario
                  required: true
                  schema:
                    type: integer
                    format: int32

            responses:
                200:
                    description: Si la peticion es correcta
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDTO'
                400:
                    description: Si el endpoint no existe
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                404:
                    description: "Si la peticon falla"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                401:
                    description: "Si el usuario no tiene autorizacion"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/UnAuthorizeRequestDTO'
                500:
                    description: "Error intreno del servidor"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/InternalErrorDTO' 
    /users/findUsers/{attribute}/{value}:
        get:
            tags:
                - User
            security:
                - bearerAuth: []
            summary: retrive a user with different value
            description: retrive a user with different value, the authentication is necesary
            operationId: retriveAttributeUser
            parameters:
                - name: attribute
                  in: path
                  required: true
                  description: Tipo de atributo a buscar
                  example: nombre
                  schema:
                    type: integer
                    format: int32
                - name: value
                  in: path
                  description: valor necesario para la busqueda
                  example: bautista
                  required: true
                  schema:
                    type: integer
                    format: int32

            responses:
                200:
                    description: "Si la peticon es correcta"
                    content:
                        application/json:
                            schema:
                                oneOf:
                                    - $ref: '#/components/schemas/UserDTO'
                                    - $ref: '#/components/schemas/UserListDTO'
                400:
                    description: Si el endpoint no existe
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                404:
                    description: "Si la peticon falla"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                401:
                    description: "Si el usuario no tiene autorizacion"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/UnAuthorizeRequestDTO'
                500:
                    description: "Error intreno del servidor"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/InternalErrorDTO'
    /users/userUpdate:
        put:
            tags: 
                - User
            security:
                - bearerAuth: []
            summary: Actualizar usuario
            description: Modificacion de usuarios
            operationId: updateUser
            requestBody:
                description: Parametro del body
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserDTO'
            responses:
                200:
                    description: Si la peticion es correcta
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDTO'
                400:
                    description: Si el endpoint no existe
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                404:
                    description: "Si la peticon falla"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                401:
                    description: "Si el usuario no tiene autorizacion"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/UnAuthorizeRequestDTO'
                500:
                    description: "Error intreno del servidor"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/InternalErrorDTO'
    /users/removeUser/{userId}:
        delete:
            tags:
                - User
            security:
                - bearerAuth: []
            summary: Remueve un usuario
            description: Elimina un usuario del sistema
            operationId: removeUser
            parameters:
                - name: userId 
                  in: path
                  description: el ID del usuario
                  required: true
                  schema:
                    type: integer
                    format: int32

            responses:
                200:
                    description: Si la peticion es correcta
                    content:
                        application/json:
                            schema:
                                message:
                                    type: string
                                    format: text
                                    default: Usuario Eliminado
                400:
                    description: Si el endpoint no existe
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                404:
                    description: "Si la peticon falla"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                401:
                    description: "Si el usuario no tiene autorizacion"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/UnAuthorizeRequestDTO'
                500:
                    description: "Error intreno del servidor"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/InternalErrorDTO'
    /users/removeUser/removeByUsername:
        delete:
            tags:
                - User
            security:
                - bearerAuth: [ ]
            summary: Remueve un usuario
            description: Elimina un usuario del sistema
            operationId: removeUserByUsername
            requestBody:
                description: Parametros de uso
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserDTO'

            responses:
                200:
                    description: Si la peticion es correcta
                    content:
                        application/json:
                            schema:
                                message:
                                    type: string
                                    format: text
                                    default: Usuario Eliminado
                400:
                    description: Si el endpoint no existe
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                404:
                    description: "Si la peticon falla"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/NotFoundMessageDTO'
                401:
                    description: "Si el usuario no tiene autorizacion"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/UnAuthorizeRequestDTO'
                500:
                    description: "Error intreno del servidor"
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/InternalErrorDTO'
...